Implementa menu de ordens de serviços

+++ src/model/services/OrdemServicoService.java
@@ -7,6 +7,7 @@ import model.entities.OrdemServico;
 import model.entities.StatusOrdem;
 import model.entities.Veiculo;
 
+import java.time.LocalDate;
 import java.util.List;
 
 public class OrdemServicoService {
@@ -15,8 +16,18 @@ public class OrdemServicoService {
 
     public void abrirOrdem(OrdemServico ordemServico) {
 
+        LocalDate agora = LocalDate.now();
+
         if(ordemServico == null) {
-            throw new RuntimeException("Ordem de serviço inválida!");
+            throw new IllegalStateException("Ordem de serviço inválida!");
+        }
+
+        if(ordemServico.getDataSaida().isBefore(ordemServico.getDataEntrada())) {
+            throw new IllegalStateException("Não é possível sair antes da entrada!");
+        }
+
+        if(ordemServico.getValor() <= 0) {
+            throw new IllegalStateException("Valor inválido!");
         }
 
         ordemServicoDao.insert(ordemServico);
@@ -24,8 +35,18 @@ public class OrdemServicoService {
 
     public void atualizarOrdemServico(OrdemServico ordemServico) {
 
+        LocalDate agora = LocalDate.now();
+
         if(ordemServico == null) {
-            throw new RuntimeException("Ordem de serviço inválida!");
+            throw new IllegalStateException("Ordem de serviço inválida!");
+        }
+
+        if(ordemServico.getDataSaida().isBefore(ordemServico.getDataEntrada())) {
+            throw new IllegalStateException("Não é possível sair antes da entrada!");
+        }
+
+        if(ordemServico.getValor() <= 0) {
+            throw new IllegalStateException("Valor inválido!");
         }
 
         ordemServicoDao.update(ordemServico);
@@ -33,8 +54,12 @@ public class OrdemServicoService {
 
     public void deletarOrdemServico(Integer id) {
 
+        if(id == null) {
+            throw new IllegalStateException("Id inválido!");
+        }
+
         if(ordemServicoDao.findById(id) == null) {
-            throw new RuntimeException("Ordem de serviço não existe!");
+            throw new IllegalStateException("Ordem de serviço não existe!");
         }
 
         ordemServicoDao.deleteById(id);
@@ -43,7 +68,7 @@ public class OrdemServicoService {
     public OrdemServico buscarOrdemServico(Integer id) {
 
         if(id == null) {
-            throw new RuntimeException("Id inválido!");
+            throw new IllegalStateException("Id inválido!");
         }
 
         return ordemServicoDao.findById(id);
@@ -53,7 +78,7 @@ public class OrdemServicoService {
     public List<OrdemServico> buscarOrdemPorVeiculo(Veiculo veiculo) {
 
         if(veiculo == null) {
-            throw new RuntimeException("Veículo inválido!");
+            throw new IllegalStateException("Veículo inválido!");
         }
 
         return ordemServicoDao.findByVeiculo(veiculo);
@@ -72,7 +97,7 @@ public class OrdemServicoService {
     public List<OrdemServico> buscarOrdemPorStatus(StatusOrdem status) {
 
         if(status == null) {
-            throw new RuntimeException("Status vazio!");
+            throw new IllegalStateException("Status vazio!");
         }
 
         return ordemServicoDao.findByStatus(status);
